Remove deployment reconciler
